pulled scatter plot state updater into own class
