+++ CoreScatterPlotView/src/au/gov/asd/tac/constellation/views/scatterplot/ScatterOptionsPane.java
@@ -18,23 +18,16 @@ package au.gov.asd.tac.constellation.views.scatterplot;
 import au.gov.asd.tac.constellation.graph.Attribute;
 import au.gov.asd.tac.constellation.graph.GraphAttribute;
 import au.gov.asd.tac.constellation.graph.GraphElementType;
-import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
 import au.gov.asd.tac.constellation.graph.ReadableGraph;
-import au.gov.asd.tac.constellation.plugins.PluginException;
 import au.gov.asd.tac.constellation.plugins.PluginExecution;
-import au.gov.asd.tac.constellation.plugins.PluginInfo;
-import au.gov.asd.tac.constellation.plugins.PluginInteraction;
-import au.gov.asd.tac.constellation.plugins.PluginType;
-import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
-import au.gov.asd.tac.constellation.plugins.templates.SimpleEditPlugin;
 import au.gov.asd.tac.constellation.utilities.color.ConstellationColor;
 import au.gov.asd.tac.constellation.utilities.icon.UserInterfaceIconProvider;
 import au.gov.asd.tac.constellation.views.scatterplot.axis.AxisBuilder;
 import au.gov.asd.tac.constellation.views.scatterplot.axis.CategoryAxisBuilder;
 import au.gov.asd.tac.constellation.views.scatterplot.axis.LogarithmicAxisBuilder;
 import au.gov.asd.tac.constellation.views.scatterplot.axis.NumberAxisBuilder;
-import au.gov.asd.tac.constellation.views.scatterplot.state.ScatterPlotConcept;
 import au.gov.asd.tac.constellation.views.scatterplot.state.ScatterPlotState;
+import au.gov.asd.tac.constellation.views.scatterplot.state.ScatterPlotStateWriter;
 import java.util.HashMap;
 import java.util.Map;
 import javafx.application.Platform;
@@ -339,35 +332,5 @@ public class ScatterOptionsPane extends BorderPane {
             optionsToolBar.setDisable(false);
             optionsPane.setDisable(false);
         });
-    }
-
-    /**
-     * Write the given ScatterPlotState to the active graph.
-     */
-    @PluginInfo(pluginType = PluginType.UPDATE, tags = {"LOW LEVEL"})
-    private static class ScatterPlotStateWriter extends SimpleEditPlugin {
-
-        private final ScatterPlotState scatterPlotState;
-
-        public ScatterPlotStateWriter(ScatterPlotState scatterPlotState) {
-            this.scatterPlotState = scatterPlotState;
-        }
-
-        @Override
-        public void edit(final GraphWriteMethods graph, final PluginInteraction interaction, final PluginParameters parameters) throws InterruptedException, PluginException {
-            final int scatterPlotStateAttribute = ScatterPlotConcept.MetaAttribute.SCATTER_PLOT_STATE.ensure(graph);
-            final ScatterPlotState state = new ScatterPlotState(scatterPlotState);
-            graph.setObjectValue(scatterPlotStateAttribute, 0, state);
-        }
-
-        @Override
-        protected boolean isSignificant() {
-            return true;
-        }
-
-        @Override
-        public String getName() {
-            return "Scatter Plot: Update State";
-        }
     }
 }
